+++ source/SIL.AppBuilder.Portal.Frontend/src/lib/auth/get-permissions.ts
