+++ source/SIL.AppBuilder.Portal.Frontend/src/lib/auth/index.tsx
@@ -1,4 +1,4 @@
 export { requireNoAuth } from './require-no-auth';
 export { requireAuth } from './require-auth';
 export { storePath, retrievePath } from './return-to';
-export { getPermissions } from './get-current-permissions';
+export { getPermissions } from './get-permissions';
